requirements.txt
astroid==2.12.12
certifi==2022.9.24
charset-normalizer==2.1.1
dill==0.3.6
idna==3.4
isort==5.10.1
jsonpickle==2.2.0
lazy-object-proxy==1.8.0
mccabe==0.7.0
platformdirs==2.5.3
python-digitalocean==1.17.0
requests==2.28.1
tomli==2.0.1
tomlkit==0.11.6
urllib3==1.26.12
wrapt==1.14.1
